# Updated feature importance defaults